move regulate logics into mixin
